replaced requirement clld with bibtexparser
The only thing we required from clld was a bibtex parser; but the price
we paid was pulling in lots of other packages, e.g. pyramid.
So using bibtexparser makes the installation a lot leaner. But again
this comes at a price, namely we have to stick to standard bibtex entry
types - which may be good practice anyway.